requirements.txt
flask
flask-pymongo
pytesseract
pillow
transformers
torch
flasgger
pdf2image